+++ bevy_rapier2d/examples/player_movement2.rs
@@ -24,8 +24,11 @@ fn main() {
 #[derive(Component)]
 pub struct Player(f32);
 
-pub fn spawn_player(mut commands: Commands, mut rapier_config: Query<&mut RapierConfiguration>) {
-    let mut rapier_config = rapier_config.single_mut();
+pub fn spawn_player(
+    mut commands: Commands,
+    mut rapier_config: Query<&mut RapierConfiguration>,
+) -> Result<()> {
+    let mut rapier_config = rapier_config.single_mut()?;
     // Set gravity to 0.0 and spawn camera.
     rapier_config.gravity = Vec2::ZERO;
     commands.spawn(Camera2d);
@@ -44,6 +47,8 @@ pub fn spawn_player(mut commands: Commands, mut rapier_config: Query<&mut Rapier
         Collider::ball(sprite_size / 2.0),
         Player(100.0),
     ));
+
+    Ok(())
 }
 
 pub fn player_movement(