+++ bevy_rapier3d/examples/ray_casting3.rs
@@ -74,23 +74,23 @@ pub fn setup_physics(mut commands: Commands) {
 
 pub fn cast_ray(
     mut commands: Commands,
-    windows: Query<&Window, With<PrimaryWindow>>,
+    window: Single<&Window, With<PrimaryWindow>>,
     rapier_context: ReadRapierContext,
     cameras: Query<(&Camera, &GlobalTransform)>,
-) {
-    let window = windows.single();
-
+) -> Result<()> {
     let Some(cursor_position) = window.cursor_position() else {
-        return;
+        log::warn!("Cursor is outside the window area");
+        return Ok(());
     };
 
     // We will color in read the colliders hovered by the mouse.
     for (camera, camera_transform) in &cameras {
         // First, compute a ray from the mouse position.
         let Ok(ray) = camera.viewport_to_world(camera_transform, cursor_position) else {
-            return;
+            log::warn!("Unable to compute ray from mouse position");
+            return Ok(());
         };
-        let context = rapier_context.single();
+        let context = rapier_context.single()?;
         // Then cast the ray.
         let hit = context.cast_ray(
             ray.origin,
@@ -108,4 +108,6 @@ pub fn cast_ray(
             commands.entity(entity).insert(ColliderDebugColor(color));
         }
     }
+
+    Ok(())
 }