+++ bevy_rapier_benches3d/src/common.rs
@@ -6,7 +6,7 @@ pub fn default_app() -> App {
 
     app.add_plugins((
         MinimalPlugins,
-        HierarchyPlugin,
+        TransformPlugin,
         TransformPlugin,
         RapierPhysicsPlugin::<()>::default(),
     ));