+++ src/plugin/systems/rigid_body.rs
@@ -14,7 +14,7 @@ use std::collections::HashMap;
 pub type RigidBodyWritebackComponents<'a> = (
     &'a RapierRigidBodyHandle,
     &'a RapierContextEntityLink,
-    Option<&'a Parent>,
+    Option<&'a ChildOf>,
     Option<&'a mut Transform>,
     Option<&'a mut TransformInterpolation>,
     Option<&'a mut Velocity>,
@@ -285,7 +285,7 @@ pub fn apply_rigid_body_user_changes(
                 }
             }
 
-            mass_modified.send(entity.into());
+            mass_modified.write(entity.into());
         }
     }
 
@@ -411,7 +411,7 @@ pub fn writeback_rigid_bodies(
         (With<RigidBody>, Without<RigidBodyDisabled>),
     >,
 ) {
-    for (handle, link, parent, transform, mut interpolation, mut velocity, mut sleeping) in
+    for (handle, link, child_of, transform, mut interpolation, mut velocity, mut sleeping) in
         writeback.iter_mut()
     {
         let config = config
@@ -465,7 +465,7 @@ pub fn writeback_rigid_bodies(
                 //       that we do to detect if the user’s transform has to be written
                 //       into the rigid-body.
                 if let Some(parent_global_transform) =
-                    parent.and_then(|p| global_transforms.get(**p).ok())
+                    child_of.and_then(|c| global_transforms.get(c.parent()).ok())
                 {
                     // We need to compute the new local transform such that:
                     // curr_parent_global_transform * new_transform = interpolated_pos
@@ -661,7 +661,7 @@ pub fn init_rigid_bodies(
         // Get rapier context from RapierContextEntityLink or insert its default value.
         let context_entity = entity_context_link.map_or_else(
             || {
-                let context_entity = default_context_access.get_single().ok()?;
+                let context_entity = default_context_access.single().ok()?;
                 commands
                     .entity(entity)
                     .insert(RapierContextEntityLink(context_entity));