+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 ## Added
 
 - Added a serialization `serialization2` example for `bevy_rapier2d`.
-- Added reflection for `Default` in addition to `Component` [#649](https://github.com/dimforge/bevy_rapier/pull/649)
+- Added reflection for `Default` in addition to `Component`. [#649](https://github.com/dimforge/bevy_rapier/pull/649)
 
 ## Modified
 
+- Update Bevy to `0.16`.
 - Update from rapier `0.23` to rapier `0.24`,
   see [rapier's changelog](https://github.com/dimforge/rapier/blob/master/CHANGELOG.md).
 - `RapierContextInitialization::InitializeDefaultRapierContext` now has more fields for better control over default physics context.